dom/media/test/test_mediastream_as_eventarget.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test MediaStream as EventTarget</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    var ms = new MediaStream();
    var listenerForFooWasCalled = false;
    ms.addEventListener("foo", () => { listenerForFooWasCalled = true; });
    var listenerForAddTrackWasCalled = false;
    ms.addEventListener("addtrack", () => { listenerForAddTrackWasCalled = true; });
    var handlerForAddTrackWasCalled = false;
    ms.onaddtrack = () => { handlerForAddTrackWasCalled = true; };

    ms.dispatchEvent(new Event("foo"));
    ms.dispatchEvent(new Event("addtrack"));

    ok(listenerForFooWasCalled,
       "Should have called the event listener for 'foo'");
    ok(listenerForAddTrackWasCalled,
       "Should have called the event listener for 'addtrack'");
    ok(handlerForAddTrackWasCalled,
       "Should have called the event handler for 'addtrack'")
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
